requirements.txt
psycopg2>=2.7.5
django>=2.1
djangorestframework>=3.8.2
django-cors-headers==2.1.0
pandas>=0.23
xlrd>= 0.9.0